Allow empty profile directories
For updating and listing profiles, we try to filter out profiles that do
not look like they were installed by Trollauncher, to avoid clobbering
other random profiles. This changes the detection so empty directories
are detected as potential profiles.

Probably not a very useful or commonly used feature, but seemed like a
good thing to have anyway.